test: move CamundaModdle tests to single file
