+++ test/spec/CamundaModdleExtensionSpec.js
